404GP-4 Change StarBar to show half stars on hover and select

+++ 404_game_portal.backend/Controllers/RatingController.cs
@@ -13,7 +13,7 @@ public class RatingController(IGameRatingRepository gameRatingRepository, IUserR
     : ControllerBase
 {
     [HttpPost]
-    public async Task<RatingViewModel> Rate([FromQuery] int rating, [FromQuery] Guid gameId)
+    public async Task<RatingViewModel> Rate([FromQuery] double rating, [FromQuery] Guid gameId)
     {
         var creationViewModel = new RatingCreationViewModel
         {
